requirements.txt
aiofiles
typing_extensions
httpx
loguru
beautifulsoup4
lxml
openai
python-dotenv
pydantic
pydantic_settings
requests
sqlalchemy
scikit-learn
markdownify
tqdm
rich
prompt_toolkit>=3.0
pyyaml
mcp
tiktoken
puremagic
qdrant-client
langfuse
python-frontmatter